Config file/gurvan

+++ Server/Source/System/NewPlayerSystem/NewPlayerSystem.cpp
@@ -27,12 +27,12 @@ void NewPlayerSystem::operator()(Registry &registry, Sparse_array<component::cne
 
         if (position[spaceShip] && type[spaceShip]) {
             std::cout <<"[SERVER] adding new player" << std::endl;
-            netqueue[0].value().toSendNetworkQueue.push(serialize_header::serializeHeader<packet_new_entity>(NETWORK_SERVER_TO_CLIENT::PACKET_TYPE::NEW_PLAYER, {spaceShip, position[spaceShip].value().x, position[spaceShip].value().y, 1, static_cast<uint16_t>(type[spaceShip].value().type)}));
+            netqueue[0].value().toSendNetworkQueue.push(serialize_header::serializeHeader<packet_new_entity>(NETWORK_SERVER_TO_CLIENT::PACKET_TYPE::NEW_PLAYER, {spaceShip, position[spaceShip].value().x, position[spaceShip].value().y, 1, type[spaceShip].value().type}));
             for (std::size_t index = 1; index != position.size(); index++) {
                 if (position[index] && type[index]) {
                     if (index == spaceShip)
                         continue;
-                    netqueue[0].value().toSendNetworkQueue.push(serialize_header::serializeHeader<packet_new_entity>(NETWORK_SERVER_TO_CLIENT::PACKET_TYPE::NEW_ENTITY, {(Entity)index, position[index].value().x, position[index].value().y, 1, static_cast<uint16_t>(type[index].value().type)}));
+                    netqueue[0].value().toSendNetworkQueue.push(serialize_header::serializeHeader<packet_new_entity>(NETWORK_SERVER_TO_CLIENT::PACKET_TYPE::NEW_ENTITY, {(Entity)index, position[index].value().x, position[index].value().y, 1, type[index].value().type}));
                 }
             }
         }
